Update lists.cpp
Unlinking nodes in link lists didn't reset the next pointers to NULL. This became an issue when adding a "recycled" node back into the top of a linked list and it went bonkers. All the "link in" methods expected new nodes that had the next pointers preset to NULL. Now inlinking a node resets the next pointers to NULL as they should.