Added heapStr() & freeStr()
heapStr(char** resultStr,const char* inStr)

This allocates room in your result string then copies your inputted string to it. This way all the check if null, check size, allocate memory, copy chars.. Is done in one fell swoop! AND since it's based on resizeBuff() you can use it over and over for different strings an it handles all the allocation and freeing of the memory. For a final freeing of the RAM freeStr() is included as well.